+++ scripts/Api.js
@@ -1,4 +1,4 @@
-class Api {
+export class Api {
   // Можно лучше: В качестве параметров при создании можно передать, например, базовый url сервера и авторизационный токен.
   // Которые можно использовать в каждом запросе.
   /*constructor() {
