+++ scripts/Card.js
@@ -1,4 +1,4 @@
-class Card {
+export class Card {
   constructor(name, link, container) {
     this.name = name;
     this.link = link;
