+++ src/index.html
@@ -5,12 +5,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <title>Mesto</title>
-    <link rel="stylesheet" href="./pages/index.css" />
+    <link rel="stylesheet" href='<%=htmlWebpackPlugin.files.css %>'>
   </head>
   <body>
     <div class="root">
       <header class="header root__section">
-        <img src="./images/logo.svg" alt="mesto logo" class="logo" />
+        <img src="<%=require('../images/logo.svg')%>" alt="mesto logo" class="logo" />
       </header>
       <div class="profile root__section">
         <div class="user-info">
@@ -26,7 +26,7 @@
       <div class="places-list root__section"></div>
       <div class="popup">
         <div class="popup__content">
-          <img src="./images/close.svg" alt="" class="popup__close" />
+          <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close" />
           <h3 class="popup__title">Новое место</h3>
           <form class="popup__form" name="new">
             <input
@@ -48,7 +48,7 @@
       <div class="popup-edit edit">
         <div class="popup__content">
           <img
-            src="./images/close.svg"
+            src="<%=require('../images/close.svg')%>"
             alt=""
             class="popup__close popup-edit__close"
           />
@@ -88,22 +88,13 @@
         <div class="popup-image__content">
           <img class="place-card__imageOpen" src="" alt="" />
           <img
-            src="./images/close.svg"
+            src="<%=require('../images/close.svg')%>"
             alt=""
             class="popup__close popup-image__close"
           />
         </div>
       </div>
     </div>
-    <script src="./scripts/cardMassive.js"></script>
-    <script src="./scripts/Card.js"></script>
-    <script src="./scripts/CardList.js"></script>
-    <script src="./scripts/Popup.js"></script>
-    <script src="./scripts/UserInfo.js"></script>
-    <script src="./scripts/Api.js"></script>
-    <script src="./scripts/FormValidator.js"></script>
-    <script src="script.js"></script>
-
-
+    <script src="<%= htmlWebpackPlugin.files.js %>"></script>
   </body>
 </html>
