+++ webpack.config.js
@@ -1,0 +1,100 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const WebpackMd5Hash = require('webpack-md5-hash');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin"); 
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const isDev = process.env.NODE_ENV === 'development';
+const webpack = require('webpack');
+
+module.exports = {
+  entry: { main: './src/index.js' },
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: '[name].[chunkhash].js'
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader"
+        }
+      },
+
+      {
+        test: /\.css$/i,
+        use: [
+                        (isDev ? 'style-loader' : MiniCssExtractPlugin.loader),
+                        'css-loader', 
+                        'postcss-loader'
+                ]
+},
+ {
+  test: /\.(png|jpg|gif|ico|svg)$/i,
+  use: [
+    {
+      loader: 'file-loader',
+      options: {
+        name: '[name].[ext]',
+        publicPath: 'img',
+        outputPath: 'img',
+        useRelativePath: true,
+        esModule: false,
+      }
+    },
+
+    {
+      loader: 'image-webpack-loader',
+      options: {
+
+      }
+    }
+  ]
+},
+ /*     {
+        test: /\.(png|jpg|gif|ico|svg)$/,
+        use: [
+                'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+                {
+                        loader: 'image-webpack-loader',
+                        options: {
+                        }
+                },
+        ]
+ },*/
+ {
+    test: /\.(eot|ttf|woff|woff2)$/,
+    loader: 'file-loader?name=./vendor/[name].[ext]'
+}/*,
+      {
+        test: /\.css$/,
+        use:  [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'] // добавили минификацию CSS
+      }*/
+    ]
+  },
+  plugins: [
+
+    new MiniCssExtractPlugin({ 
+      filename: 'style.[contenthash].css',
+    }),
+    new OptimizeCssAssetsPlugin({
+        assetNameRegExp: /\.css$/g,
+        cssProcessor: require('cssnano'),
+        cssProcessorPluginOptions: {
+                preset: ['default'],
+        },
+        canPrint: true
+}),
+    new WebpackMd5Hash(),
+    new webpack.DefinePlugin({
+        'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
+    }),
+    new HtmlWebpackPlugin({
+      inject: false,
+      hash: true,
+      template: './src/index.html',
+      filename: 'index.html'
+    })
+  ]
+};
